refactor(kubernetes): clean up infrastructure interfaces (#8450)
* refactor(kubernetes): colocate kubernetes infrastructure interfaces

The kubernetes infrastructure interfaces currently represent an amalgam of data received from clouddriver from each logical kind request and data from a separate request for the full resource manifest. There is also some overlap among these interfaces. Let's colocate them to improve readability of the upcoming refactor, at the top level because these logical kind interfaces are relevant beyond the details directories.

* refactor(core): remove redundant namespace check from cluster.service.ts

Clouddriver sets Kubernetes server groups' `region` equal to their `namespace`, so we should not need to check both ([source](https://github.com/spinnaker/clouddriver/blob/master/clouddriver-kubernetes/src/main/java/com/netflix/spinnaker/clouddriver/kubernetes/caching/view/model/ManifestBasedModel.java#L46)).

* refactor(core): remove displayName from IServerGroupManager

Clouddriver only guarantees that server group managers have a name ([source](https://github.com/spinnaker/clouddriver/blob/master/clouddriver-core/src/main/groovy/com/netflix/spinnaker/clouddriver/model/ServerGroupManager.java#L30)). The only extant implementation of this interface is the Kubernetes-specific server group manager, which does guarantee a displayName, but we should use the name instead in non-Kubernetes-specific code. The result of this from a user perspective will be to include the Kubernetes kind along with the name in these entity tag labels.

* refactor(core): remove namespace from IServerGroup

Only the Kubernetes provider uses namespaces this way; let's not define this as an optional on the core server group interface.

* refactor(kubernetes): move shared infrastructure fields to shared interface

Let's move the common Kubernetes infrastructure fields of apiVersion, displayName, namespace, and kind to an interface that each logical kind interface can extend.
