+++ app/scripts/modules/kubernetes/src/instance/index.ts
@@ -1,2 +1,1 @@
 export * from './details/details.controller';
-export * from './details/IKubernetesInstance';
