--- app/scripts/modules/kubernetes/src/loadBalancer/details/IKubernetesLoadBalancer.ts
@@ -1,8 +1,0 @@
-import { ILoadBalancer } from '@spinnaker/core';
-
-export interface IKubernetesLoadBalancer extends ILoadBalancer {
-  kind: string;
-  displayName: string;
-  apiVersion: string;
-  namespace: string;
-}
