+++ app/scripts/modules/kubernetes/src/serverGroup/index.ts
@@ -2,4 +2,3 @@ export * from './serverGroupCommandBuilder.service';
 export * from './serverGroupTransformer.service';
 export * from './details/details.controller';
 export * from './details/resize/resize.controller';
-export * from './details/IKubernetesServerGroup';
